Tensor Operations

Base.inv

LinearAlgebra.dot

LinearAlgebra.norm

LinearAlgebra.tr

Statistics.mean

Tensorial.contraction

Tensorial.dev

Tensorial.otimes

Tensorial.rotmat

Tensorial.rotmat

Tensorial.rotmat

Tensorial.skew

Tensorial.vol

Base.inv — Method

inv(::AbstractSecondOrderTensor)
inv(::AbstractSymmetricSecondOrderTensor)
inv(::AbstractFourthOrderTensor)
inv(::AbstractSymmetricFourthOrderTensor)

Compute the inverse of a tensor.

Examples

julia> x = rand(SecondOrderTensor{3})
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 0.590845  0.460085  0.200586
 0.766797  0.794026  0.298614
 0.566237  0.854147  0.246837

julia> inv(x)
3×3 Tensor{Tuple{3,3},Float64,2,9}:
  19.7146   -19.2802    7.30384
   6.73809  -10.7687    7.55198
 -68.541     81.4917  -38.8361

julia> x ⋅ inv(x) ≈ one(I)
true

LinearAlgebra.dot — Method

dot(x::AbstractTensor, y::AbstractTensor)
x ⋅ y

Compute dot product such as $a = x_i y_i$. This is equivalent to contraction(::AbstractTensor, ::AbstractTensor, Val(1)). x ⋅ y (where ⋅ can be typed by \cdot<tab>) is a synonym for cdot(x, y).

Examples

julia> x = rand(Vec{3})
3-element Tensor{Tuple{3},Float64,1,3}:
 0.5908446386657102
 0.7667970365022592
 0.5662374165061859

julia> y = rand(Vec{3})
3-element Tensor{Tuple{3},Float64,1,3}:
 0.4600853424625171
 0.7940257103317943
 0.8541465903790502

julia> a = x ⋅ y
1.3643452781654775

LinearAlgebra.norm — Method

norm(::AbstractTensor)

Compute norm of a tensor.

Examples

julia> x = rand(Mat{3, 3})
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 0.590845  0.460085  0.200586
 0.766797  0.794026  0.298614
 0.566237  0.854147  0.246837

julia> norm(x)
1.7377443667834922

LinearAlgebra.tr — Method

tr(::AbstractSecondOrderTensor)
tr(::AbstractSymmetricSecondOrderTensor)

Compute the trace of a square tensor.

Examples

julia> x = rand(Mat{3,3})
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 0.590845  0.460085  0.200586
 0.766797  0.794026  0.298614
 0.566237  0.854147  0.246837

julia> tr(x)
1.6317075356075135

Statistics.mean — Method

mean(::AbstractSecondOrderTensor)
mean(::AbstractSymmetricSecondOrderTensor)

Compute the mean value of diagonal entries of a square tensor.

Examples

julia> x = rand(Mat{3,3})
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 0.590845  0.460085  0.200586
 0.766797  0.794026  0.298614
 0.566237  0.854147  0.246837

julia> mean(x)
0.5439025118691712

Tensorial.contraction — Method

contraction(::AbstractTensor, ::AbstractTensor, ::Val{N})

Conduct contraction of N inner indices. For example, N=2 contraction for third-order tensors $A_{ij} = B_{ikl} C_{klj}$ can be computed in Tensorial.jl as

julia> B = rand(Tensor{Tuple{3,3,3}});

julia> C = rand(Tensor{Tuple{3,3,3}});

julia> A = contraction(B, C, Val(2))
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 1.36912   1.86751  1.32531
 1.61744   2.34426  1.94101
 0.929252  1.89656  1.79015

Following symbols are also available for specific contractions:

x ⊗ y (where ⊗ can be typed by \otimes<tab>): contraction(x, y, Val(0))

x ⋅ y (where ⋅ can be typed by \cdot<tab>): contraction(x, y, Val(1))

x ⊡ y (where ⊡ can be typed by \boxdot<tab>): contraction(x, y, Val(2))

Tensorial.dev — Method

dev(::AbstractSecondOrderTensor{3})
dev(::AbstractSymmetricSecondOrderTensor{3})

Compute the deviatoric part of a square tensor. Support only for tensors in 3D.

Examples

julia> x = rand(Mat{3,3})
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 0.590845  0.460085  0.200586
 0.766797  0.794026  0.298614
 0.566237  0.854147  0.246837

julia> dev(x)
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 0.0469421  0.460085   0.200586
 0.766797   0.250123   0.298614
 0.566237   0.854147  -0.297065

julia> tr(dev(x))
0.0

Tensorial.otimes — Method

otimes(x::AbstractTensor, y::AbstractTensor)
x ⊗ y

Compute tensor product such as $A_{ij} = x_i y_j$. x ⊗ y (where ⊗ can be typed by \otimes<tab>) is a synonym for otimes(x, y).

Examples

julia> x = rand(Vec{3})
3-element Tensor{Tuple{3},Float64,1,3}:
 0.5908446386657102
 0.7667970365022592
 0.5662374165061859

julia> y = rand(Vec{3})
3-element Tensor{Tuple{3},Float64,1,3}:
 0.4600853424625171
 0.7940257103317943
 0.8541465903790502

julia> A = x ⊗ y
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 0.271839  0.469146  0.504668
 0.352792  0.608857  0.654957
 0.260518  0.449607  0.48365

Tensorial.rotmat — Method

rotmat(θ::Real)

Construct 2D rotation matrix.

Examples

julia> rotmat(deg2rad(30))
2×2 Tensor{Tuple{2,2},Float64,2,4}:
 0.866025  -0.5
 0.5        0.866025

Tensorial.rotmat — Method

rotmat(θ::Vec{3}; sequence::Symbol, degree::Bool = false)
rotmatx(θ::Real)
rotmaty(θ::Real)
rotmatz(θ::Real)

Convert Euler angles to rotation matrix. Use 3 characters belonging to the set (X, Y, Z) for intrinsic rotations, or (x, y, z) for extrinsic rotations.

Examples

julia> α, β, γ = rand(Vec{3});

julia> rotmat(Vec(α,β,γ), sequence = :XYZ) ≈ rotmatx(α) ⋅ rotmaty(β) ⋅ rotmatz(γ)
true

julia> rotmat(Vec(α,β,γ), sequence = :xyz) ≈ rotmatz(γ) ⋅ rotmaty(β) ⋅ rotmatx(α)
true

julia> rotmat(Vec(α,β,γ), sequence = :XYZ) ≈ rotmat(Vec(γ,β,α), sequence = :zyx)
true

Tensorial.rotmat — Method

rotmat(a => b)

Construct rotation matrix rotating vector a to b. The norms of two vectors must be the same.

Examples

julia> a = rand(Vec{3}); a /= norm(a);

julia> b = rand(Vec{3}); b /= norm(b);

julia> R = rotmat(a => b);

julia> R ⋅ a ≈ b
true

Tensorial.skew — Method

skew(::AbstractSecondOrderTensor)
skew(::AbstractSymmetricSecondOrderTensor)

Compute skew-symmetric (anti-symmetric) part of a second order tensor.

Tensorial.vol — Method

vol(::AbstractSecondOrderTensor{3})
vol(::AbstractSymmetricSecondOrderTensor{3})

Compute the volumetric part of a square tensor. Support only for tensors in 3D.

Examples

julia> x = rand(Mat{3,3})
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 0.590845  0.460085  0.200586
 0.766797  0.794026  0.298614
 0.566237  0.854147  0.246837

julia> vol(x)
3×3 Tensor{Tuple{3,3},Float64,2,9}:
 0.543903  0.0       0.0
 0.0       0.543903  0.0
 0.0       0.0       0.543903

julia> vol(x) + dev(x) ≈ x
true