Einstein summation

Tensorial.@einsum

Tensorial.@einsum — Macro

@einsum [TensorType] (i,j...) -> expr
@einsum [TensorType] expr

Performs tensor computations using the Einstein summation convention. The arguments of the anonymous function are treated as free indices. If no arguments are provided, they are inferred based on the order in which the indices appear from left to right. Since @einsum cannot fully infer tensor symmetries, it is possible to annotate the returned tensor type (though this is not checked for correctness). This can help eliminate the computation of the symmetric part, improving performance.

Examples

julia> A = rand(Mat{3,3});

julia> B = rand(Mat{3,3});

julia> (@einsum (i,j) -> A[j,k] * B[k,i]) ≈ (A ⋅ B)'
true

julia> (@einsum A[i,k] * B[k,j]) ≈ A ⋅ B
true

julia> (@einsum A[i,j] * A[i,j]) ≈ A ⊡ A
true

julia> (@einsum SymmetricSecondOrderTensor{3} A[k,i] * A[k,j]) ≈ A' ⋅ A
true